feat: core 19 testnet fixes up to RC5 (#197)
* fix: update for Core 19 RC1

* There are new LLMQ types for Platform (TEST, DEVNET)

* fix: remove banned and expired masternodes from new quorum quarters

* this is valid on v19 and testnet

* fix: prevent NPE when closing down the systems to allow restarting without sync issues

* fix: update for Core 19 RC5

* fix: support new LLMQ type

* fix: update testnet masternode list

+++ core/src/main/java/org/bitcoinj/store/FlatDB.java
@@ -230,7 +230,7 @@ public class FlatDB<Type extends AbstractManager> {
                 }
 
                 // de-serialize file header (network specific magic number) and ..
-                pchMsgTmp = (int)Utils.readUint32(vchData, magicMessage.length());
+                pchMsgTmp = Utils.readUint32(vchData, magicMessage.length());
 
                 // ... verify the network matches ours
                 if (pchMsgTmp != context.getParams().getPacketMagic()) {
